pyproject.toml
[tool.poetry]
name = "vector-store-in-memory"
version = "0.1.0"
description = ""
authors = ["RutamBhagat <rutambhagat@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.1.16"
pypdf = "^4.2.0"
openai = "^1.23.2"
black = "^24.4.0"
python-dotenv = "^1.0.1"
tiktoken = "^0.6.0"
faiss-cpu = "^1.8.0"
langchain-openai = "^0.1.3"
chromadb = "^0.4.24"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
